requirements.txt
dj-rest-auth[with_social]==2.2.5
Django==4.1.1
django-countries==7.3.2
django-filter==22.1
djangorestframework==3.13.1
Pillow==9.2.0

# openapi
uritemplate~=4.1.1
PyYAML~=6.0
